Remove binary files
Stop tracking the SQLite database in git, because it's impossible to
resolve merge conflicts. Open the file in an editor - what would you do
if you and I both added records at the same time? There's no way to
merge the changes together.

Python bytecode is automatically generated when the files are run, so
there's no need to check them in. The bytecode is also different on
different operating systems and Python versions, so checking it in is
more trouble than its worth.